+++ scripts/deployments.json
@@ -32,6 +32,7 @@
 	"kovan": {
 		"Airdrop": "0xD9ceC0F5b2B43C74DFA4A87117968808b95d9018",
 		"OngoingAirdrop": "0xd9112fC9396d38D024D1369e72f039d542000f80",
+		"ProxysUSD": "0x57Ab1ec28D129707052df4dF418D58a2D46d5f51",
 		"VestingEscrow": "0xaF05693d2c336b40055F6531556aB6D0Ebf003D8",
 		"EscrowThales": "0xcE9c1aAb41b18b24355138776186484B260B9cf0",
 		"StakingThales": "0x8d178B3Dd7B29dfB38f55FdFE6518cabe6B31059",
@@ -46,9 +47,13 @@
 		"BinaryOptionMarketData": "0x42Af0890d254387110555A46d261B9e4f6E327Ac",
 		"OpThales_L1": "0x008e86b2e21ce8d71C4700A9a174D151839cFbc7",
 		"OwnedUpgradeabilityProxy": "0x7596C9aA9ce105610324AcC27D943C5235518701",
-		"ProxyThalesExchanger": "0x7596C9aA9ce105610324AcC27D943C5235518701",
-		"ProxyStakingThales": "0xB6C1cE599EB43daBA7A2ef90C1517d78dF821711",
-		"ProxyEscrowThales": "0x2f07803C8D59889f6E02D7ABfa8Cc7C7c571B3CA"
+		"L1StandardBridge": "0x22F24361D548e5FaAfb36d1437839f080363982B",
+		"ProxyThalesExchanger": "0xC5c9F092123758E0c6B19caadff2cE8B9Ac2360D",
+		"ProxyStakingThales": "0xc517D5A81Be4489282cDe30A77Fc1Ad74f6D8AeD",
+		"ProxyEscrowThales": "0x3Ea294aee4c588A0DCf5a155bDBB1674431c3922",
+		"ProxyThalesExchangerImplementation": "0x1687fDC8A409Be442e2dD3DE42b1fcC974Ac9eE8",
+		"ProxyStakingThalesImplementation": "0x062106C253029A0065c3829dB00f6483444510C9",
+		"ProxyEscrowThalesImplementation": "0x596C785d91A18CC6f25861770f5238d176946B53"
 	},
 	"goerli": {
 		"Airdrop": "",
@@ -90,8 +95,10 @@
 		"PriceFeed": "0x34fb310296b441625469395ed5765296F6426636",
 		"ThalesRoyale": "0x9EE65877dE8658e95f5c893581a76CCd5bc6d35B",
 		"OpThales_L2": "0x6002a8F2A8dC0B704080765cE96aac63D5f7B224",
-		"ProxyStakingThales": "0xeEAacf51A6A3009Fc0FFB1DE42B01CfA8aDF8393",
-		"ProxyEscrowThales": "0xd86aA455cc860864b51261Ee273e7298EABaeF0b"
+		"ProxyStakingThales": "0x3eD7fba8d640B2ED4E0a57607185FaB0f228515C",
+		"ProxyEscrowThales": "0x1A16ebD0a9855A9537aa3bc30989634ec3043Df2",
+		"ProxyStakingThalesImplementation": "0xEBB06f3ca3980d71EcccC99Fd7CA1080bF82Becd",
+		"ProxyEscrowThalesImplementation": "0x4f8A0ca4af3e58EB992ac2D3573065117CDB5CA0"
 	},
 	"localhost": {
 		"Airdrop": "",
