+++ scripts/l1_l2_exchanger/development_manual_deploy/deploy_OpThales_L1_L2.js
@@ -1,11 +1,10 @@
 const path = require('path');
-const { ethers, upgrades  } = require('hardhat');
+const { ethers } = require('hardhat');
 const w3utils = require('web3-utils');
+const snx = require('synthetix-2.50.4-ovm');
 const { artifacts, contract, web3 } = require('hardhat');
-require('dotenv').config();
 
-// const user_key = process.env.PRIVATE_KEY;
-const user_key = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80';
+const user_key = process.env.PRIVATE_KEY;
 
 const {
 	fastForward,
@@ -13,60 +12,51 @@ const {
 	currentTime,
 	multiplyDecimalRound,
 	divideDecimalRound,
-} = require('../../test/utils/index')();
+} = require('../../../test/utils/index')();
 
 const ZERO_ADDRESS = '0x' + '0'.repeat(40);
 
 const L2_BRIDGE_ADDRESS = '0x4200000000000000000000000000000000000010';
 
-const { getTargetAddress, setTargetAddress } = require('../helpers');
+const { getTargetAddress, setTargetAddress } = require('../../helpers');
 
 const L2StandardBridgeArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L2StandardBridge.sol/OVM_L2StandardBridge');
 const L1StandardBridgeArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L1StandardBridge.sol/OVM_L1StandardBridge');
-const { toBytes32 } = require('../../index');
+const { toBytes32 } = require('../../../index');
 
 async function main() {
 	let accounts = await ethers.getSigners();
 	let owner = accounts[0];
+	// let networkObj = await ethers.provider.getNetwork();
 	
-	const local_L1 = 'localhost'
+	let network_kovan = new ethers.providers.InfuraProvider("kovan");
+	// console.log(network_kovan);
+	const net_kovan = 'kovan'
+	let network_optimistic_kovan = await ethers.provider.getNetwork();
+	const net_optimistic_kovan = 'optimisticKovan'
+		
+	const l1Wallet = new ethers.Wallet(user_key, network_kovan);
+	const l2Wallet = new ethers.Wallet(user_key, ethers.provider);
 	
-	const local_L2 = 'optimistic'
-	
-    const l1RpcProvider = new ethers.providers.JsonRpcProvider(process.env.LOCAL_OPT_IP_2)
-    const l2RpcProvider = new ethers.providers.JsonRpcProvider(process.env.LOCAL_OPT_IP)
-	
-	const l1Wallet = new ethers.Wallet(user_key, l1RpcProvider);
-	const l2Wallet = new ethers.Wallet(user_key, l2RpcProvider);
-
-	let blockNumber = await l1RpcProvider.getBlockNumber();
+	let blockNumber = await network_kovan.getBlockNumber();
 	console.log("Kovan block number: ", blockNumber);
 	
-	blockNumber = await l2RpcProvider.getBlockNumber();
+	blockNumber = await ethers.provider.getBlockNumber();
 	console.log("Optimistic Kovan block number: ", blockNumber);
 	
+	// const L2StandardBridge = await ethers.getContractFactory('../../node_modules/@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L2StandardBridge.sol:OVM_L2StandardBridge');
 	const L2StandardBridge = new ethers.ContractFactory(L2StandardBridgeArtifacts.abi, L2StandardBridgeArtifacts.bytecode);
 	const L1StandardBridge = new ethers.ContractFactory(L1StandardBridgeArtifacts.abi, L1StandardBridgeArtifacts.bytecode);
 	const OP_Thales_L1 = await ethers.getContractFactory('/contracts/Token/OpThales_L1.sol:OpThales');
-	const ThalesExchanger = await ethers.getContractFactory('ProxyThalesExchanger');
+	const ThalesExchanger = await ethers.getContractFactory('ThalesExchanger');
 	const OP_Thales_L2 = await ethers.getContractFactory('/contracts/Token/OpThales_L2.sol:OpThales');
-
-	const OwnedUpgradeabilityProxy = await ethers.getContractFactory('OwnedUpgradeabilityProxy');
-	const OwnedUpgradeabilityProxy_deployed = await OwnedUpgradeabilityProxy.deploy();
-	await OwnedUpgradeabilityProxy_deployed.deployed();
-
-	console.log("Owned proxy deployed on: ",OwnedUpgradeabilityProxy_deployed.address);
-	setTargetAddress('OwnedUpgradeabilityProxy', local_L1, OwnedUpgradeabilityProxy_deployed.address);
+	const ThalesAddress = getTargetAddress('Thales', 'kovan');
 
 	const Thales = await ethers.getContractFactory('Thales');
-	// console.log("L2 Contract:\n", OP_Thales_L2);
-	
+		
 
-	const Thales_deployed= await Thales.connect(l1Wallet).deploy();
-    await Thales_deployed.deployed();
-    console.log("Thales deployed on: ",Thales_deployed.address);
-	setTargetAddress('Thales', local_L1, Thales_deployed.address);
-
+	const Thales_deployed= await Thales.connect(l1Wallet).attach(ThalesAddress);
+	console.log("Thales on Kovan at: ", Thales_deployed.address);
 	const ThalesExchanger_connected = await ThalesExchanger.connect(l1Wallet);
 	console.log("Thales Exchanger ready to deploy: ", ThalesExchanger_connected.signer._isSigner);
 	const OP_Thales_L1_connected = await OP_Thales_L1.connect(l1Wallet);
@@ -88,7 +78,7 @@ async function main() {
 	let tx = await OP_Thales_L1_deployed.deployed();
 	// console.log(tx);
 	console.log("Optimistic Thales L1 deployed on: ",OP_Thales_L1_deployed.address);
-	setTargetAddress('OpThales_L1', local_L1, OP_Thales_L1_deployed.address);
+	setTargetAddress('OpThales_L1', net_kovan, OP_Thales_L1_deployed.address);
 	
 	const OP_Thales_L2_deployed = await OP_Thales_L2_connected.deploy(
 		L2_BRIDGE_ADDRESS,
@@ -99,45 +89,8 @@ async function main() {
 		
 	tx = await OP_Thales_L2_deployed.deployed();
 	console.log("Optimistic Thales L2 deployed on: ",OP_Thales_L2_deployed.address);
-	setTargetAddress('OpThales_L2', local_L2, OP_Thales_L2_deployed.address);
-	
-	const ThalesExchanger_deployed = await ThalesExchanger_connected.deploy();
-	await ThalesExchanger_deployed.deployed();
+	setTargetAddress('OpThales_L2', net_optimistic_kovan, OP_Thales_L2_deployed.address);
 
-	tx = await ThalesExchanger_deployed.initialize(
-	// const ThalesExchanger_deployed = await upgrades.deployProxy(ThalesExchanger, [
-		owner.address, 
-		Thales_deployed.address,
-		OP_Thales_L1_deployed.address,
-		L1StandardBridge_deployed.address,
-		OP_Thales_L2_deployed.address
-	// ]
-	);
-
-
-	
-	// tx = await ThalesExchanger_deployed.deployed();
-	console.log("Thales Exchanger deployed on: ", ThalesExchanger_deployed.address);
-	setTargetAddress('ProxyThalesExchanger', local_L1, ThalesExchanger_deployed.address);
-	
-	
-
-	tx = await ThalesExchanger_deployed.owner.call();
-	console.log("Owner is:", tx);
-
-	tx = await OwnedUpgradeabilityProxy_deployed.upgradeTo(ThalesExchanger_deployed.address);
-	console.log("Proxy updated");
-	
-	await delay(20000);
-	
-	tx = await ThalesExchanger_deployed.transferOwnershipAtInit(OwnedUpgradeabilityProxy_deployed.address);
-	console.log("Owner changed to proxy");
-
-	console.log("Wait 20 seconds");
-	await delay(20000);
-
-	tx = await ThalesExchanger_deployed.owner.call();
-	console.log("Owner is:", tx);
 
 }
