+++ scripts/l1_l2_exchanger/development_manual_deploy/verify_proxy_L1_L2.js
@@ -5,7 +5,7 @@ const { artifacts, contract, web3 } = require('hardhat');
 
 
 
-const { getTargetAddress, setTargetAddress } = require('../helpers');
+const { getTargetAddress, setTargetAddress } = require('../../helpers');
 
 
 async function main() {
